+++ changelog.md
@@ -3,6 +3,7 @@
 ## New Features
 
 - integrated Lyx Support via Multimarkdown
+- Change the Default Converters via Preferences
 
 ## Bug Fixes